+++ app/Http/Controllers/Api/v1/PostFileController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Models\Post;
 use App\Models\PostFile;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class PostFileController extends Controller
 {
@@ -18,6 +19,7 @@ class PostFileController extends Controller
      */
     public function destroy(Post $post, PostFile $file)
     {
+        Gate::authorize('delete', [PostFile::class, $file, $post]);
         $file->delete();
         return response()->json([
             'message' => __('messages.post_file_deleted'),
